requirements.txt
fastapi
uvicorn
sqlmodel
requests
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart
python-dotenv
